permisos usuarios, order by y order

+++ TPE1/controllers/commentAPIController.php
@@ -3,6 +3,7 @@
 require_once "models/commentModel.php";
 require_once "views/apiView.php";
 require_once "models/bookModel.php";
+require_once "helpers/authHelper.php";
 
 class commentAPIController
 {
@@ -14,7 +15,7 @@ class commentAPIController
 
     public function __construct()
     {
-        // $this->authHelper = new AuthHelper();
+        $this->authHelper = new AuthHelper();
         $this->model = new commentModel();
         $this->view = new apiView();
         $this->modelBook = new bookModel();
@@ -23,22 +24,31 @@ class commentAPIController
 
     public function getCommentsByBook($params = null)
     {
-        // $this->authHelper->checkLoggedIn();
         $idBook = $params[":ID"];
+        $orderBy = null;
+        $order = null;
         // Chequear que exista ese libro, si no existe devolver 404. 
-        $comments =  $this->model->getCommentsByBook($idBook);
+        if (isset($_GET['orderBy'])) {
+            $orderBy = $_GET['orderBy'];
+        } else {
+            $orderBy = 'fecha_creacion';
+        }
+        if (isset($_GET['order']))
+            $order = $_GET['order'];
+        else
+            $order = 'desc';
+        $comments = $this->model->getCommentsByBook($idBook, $orderBy, $order);
         if ($comments) {
             return $this->view->response($comments, 200);
         } else {
-            return $this->view->response("No hay comentarios", 200);
+            return $this->view->response([], 200);
         }
     }
 
     public function addCommentBook($params = null)
     {
-        // $this->authHelper->checkLoggedIn();
+        $this->authHelper->isLogged();
         $body = $this->getBody();
-
         $idComment = $this->model->addCommentBook($body->comentario, $body->puntuacion, $body->id_usuario, $body->id_libro);
         if ($idComment) {
             $this->view->response("Se ha insertado correctamente", 400);
@@ -49,19 +59,25 @@ class commentAPIController
 
     public function deleteComment($params = null)
     {
-        // $this->authHelper->checkLoggedIn();
-        $idComment = $params[":ID"];
-        $comment = $this->model->getComment($idComment);
+        $isLogged = $this->authHelper->isLogged();
+        if ($isLogged) {
+            $isAdmin = $this->authHelper->isAdmin();
 
-        if ($comment) {
-            $this->model->deleteComment($idComment);
-            return $this->view->response("El comentario ha sido eliminado", 200);
-        } else {
-            return $this->view->response("El comentario no existe", 404);
+            if ($isAdmin) {
+                $idComment = $params[":ID"];
+                $comment = $this->model->getComment($idComment);
+
+                if ($comment) {
+                    $this->model->deleteComment($idComment);
+                    return $this->view->response("El comentario ha sido eliminado", 200);
+                } else {
+                    return $this->view->response("El comentario no existe", 404);
+                }
+            }
         }
     }
 
-    public function getBook($params = null)
+    function getBook($params = null)
     {
         $idBook =  $params[":ID"];
         $book = $this->modelBook->getBook($idBook);
